+++ lib/constants/pair.ts
@@ -1,0 +1,32 @@
+import type { ZaifPair } from '@/shared/types/currency'
+
+const ALL_ZAIF_PAIRS: ZaifPair[] = [
+  'ncxc_btc',
+  'cicc_jpy',
+  'jpyz_jpy',
+  'btc_jpy',
+  'xcp_btc',
+  'zaif_jpy',
+  'erc20.cms_jpy',
+  'xym_btc',
+  'mona_btc',
+  'zaif_btc',
+  'bch_btc',
+  'eth_jpy',
+  'eth_btc',
+  'bch_jpy',
+  'mosaic.cms_btc',
+  'xem_btc',
+  'fscc_btc',
+  'cicc_btc',
+  'xcp_jpy',
+  'xym_jpy',
+  'mosaic.cms_jpy',
+  'fscc_jpy',
+  'xem_jpy',
+  'erc20.cms_btc',
+  'ncxc_jpy',
+  'mona_jpy'
+]
+
+export { ALL_ZAIF_PAIRS }