+++ lib/utils/types/index.ts
@@ -1,0 +1,3 @@
+type StrictExtract<T, U extends T> = T extends U ? T : never
+
+export type { StrictExtract }