+++ CHANGELOG.md
@@ -1,16 +1,52 @@
+# [v1.0.0-beta.3](https://github.com/TomokiMiyauci/coincheck-client/compare/v1.0.0-beta.2...v1.0.0-beta.3) (2021-11-02)
+
+## ✨ New Features
+
+- [`22c5860`](https://github.com/TomokiMiyauci/coincheck-client/commit/22c5860) Improve public API for rate of symbol pair types
+
+# [v1.0.0-beta.2](https://github.com/TomokiMiyauci/coincheck-client/compare/v1.0.0-beta.1...v1.0.0-beta.2) (2021-11-02)
+
+## 🐛 Bug Fixes
+
+- [`8de76fe`](https://github.com/TomokiMiyauci/coincheck-client/commit/8de76fe) Add extenal module to dependencies
+
+# v1.0.0-beta.1 (2021-11-02)
+
+## ✨ New Features
+
+- [`8818753`](https://github.com/TomokiMiyauci/coincheck-client/commit/8818753) Change public api interface
+
 # Change Log
 
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [0.0.1-alpha.1](https://github.com/TomokiMiyauci/zaif-client/compare/v0.0.1-alpha.0...v0.0.1-alpha.1) (2021-01-12)
+# [0.1.0-alpha.4](https://github.com/TomokiMiyauci/coincheck-client/compare/v0.1.0-alpha.3...v0.1.0-alpha.4) (2020-12-02)
 
-**Note:** Version bump only for package zaif-client
+### Bug Fixes
 
+- **coincheck-client:** change to relative path ([58d8160](https://github.com/TomokiMiyauci/coincheck-client/commit/58d81600747b664a83fd524c5aa2a46d04f4e874))
 
+# [0.1.0-alpha.3](https://github.com/TomokiMiyauci/coincheck-client/compare/v0.1.0-alpha.2...v0.1.0-alpha.3) (2020-12-02)
 
+**Note:** Version bump only for package coincheck-client
 
+# [0.1.0-alpha.2](https://github.com/TomokiMiyauci/coincheck-client/compare/v0.1.0-alpha.1...v0.1.0-alpha.2) (2020-12-02)
 
-## 0.0.1-alpha.0 (2020-12-15)
+**Note:** Version bump only for package coincheck-client
 
-**Note:** Version bump only for package zaif-client
+# [0.1.0-alpha.1](https://github.com/TomokiMiyauci/coincheck-client/compare/v0.1.0-alpha.0...v0.1.0-alpha.1) (2020-12-02)
+
+**Note:** Version bump only for package coincheck-client
+
+# 0.1.0-alpha.0 (2020-12-02)
+
+### Features
+
+- **coincheck-client:** add api/exchange-orders-rate ([fbb8378](https://github.com/TomokiMiyauci/coincheck-client/commit/fbb8378c1d341f95beafef820edde90709058471))
+- **coincheck-client:** add api/order-books ([e382954](https://github.com/TomokiMiyauci/coincheck-client/commit/e382954eddbb66a97fc6601ec14f28e7f2f285f1))
+- **coincheck-client:** add api/rate ([0f6508a](https://github.com/TomokiMiyauci/coincheck-client/commit/0f6508ac683ff82c6208bedd0d7b2a117078609a))
+- **coincheck-client:** add api/ticker ([1830500](https://github.com/TomokiMiyauci/coincheck-client/commit/183050053efd6336e259866680b230b3d308e100))
+- **coincheck-client:** add api/trades ([ecb6e3d](https://github.com/TomokiMiyauci/coincheck-client/commit/ecb6e3d0e695cd35e46f26c64e77284d4e822aa4))
+- **coincheck-client:** add base instance ([c10f49f](https://github.com/TomokiMiyauci/coincheck-client/commit/c10f49ffc73d94771863f0910adfb3e6d462e408))
+- **coincheck-client:** add types of pair ([381fa8e](https://github.com/TomokiMiyauci/coincheck-client/commit/381fa8e6ce00f29f3bd1ef5280cc037d8d8f958a))